moved prompt recompilation
